pyproject.toml
[project]
name = "saqg"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "comfykit>=0.1.12",
    "deep-translator>=1.11.4",
    "diffusers>=0.36.0",
    "loguru>=0.7.3",
    "openai>=2.14.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyinstaller>=6.17.0",
    "pyside6>=6.10.1",
    "transformers>=4.57.3",
]
